refactor: remove abstraction from content services
