+++ content/src/main/java/pt/estga/content/services/MarkQueryService.java
@@ -1,20 +1,39 @@
 package pt.estga.content.services;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import pt.estga.content.entities.Mark;
+import pt.estga.content.repositories.MarkRepository;
 
 import java.util.Optional;
 
-public interface MarkQueryService {
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class MarkQueryService {
 
-    Page<Mark> findAll(Pageable pageable);
+    private final MarkRepository repository;
 
-    Page<Mark> findAllManagement(Pageable pageable);
+    public Page<Mark> findAll(Pageable pageable) {
+        return repository.findByActiveIsTrue(pageable);
+    }
 
-    Optional<Mark> findById(Long id);
+    public Page<Mark> findAllManagement(Pageable pageable) {
+        return repository.findAll(pageable);
+    }
 
-    Optional<Mark> findWithCoverById(Long id);
+    public Optional<Mark> findById(Long id) {
+        return repository.findById(id);
+    }
 
-    long count();
+    public Optional<Mark> findWithCoverById(Long id) {
+        return repository.findWithCoverById(id);
+    }
+
+    public long count() {
+        return repository.count();
+    }
 }
