+++ content/src/main/java/pt/estga/content/services/MarkService.java
@@ -1,22 +1,61 @@
 package pt.estga.content.services;
 
+import lombok.RequiredArgsConstructor;
+import org.jspecify.annotations.NonNull;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import pt.estga.content.entities.Mark;
+import pt.estga.content.events.MarkCreatedEvent;
+import pt.estga.content.repositories.MarkRepository;
 import pt.estga.file.entities.MediaFile;
 
 import java.util.Optional;
 
-public interface MarkService {
+@Service
+@RequiredArgsConstructor
+public class MarkService {
 
-    Optional<Mark> findById(Long id);
+    private final MarkRepository repository;
+    private final ApplicationEventPublisher eventPublisher;
 
-    Mark create(Mark mark);
+    public Optional<Mark> findById(Long id) {
+        return repository.findById(id);
+    }
 
-    Mark create(Mark mark, MediaFile cover);
+    @Transactional
+    public Mark create(Mark mark) {
+        return create(mark, null);
+    }
 
-    Mark update(Mark mark);
+    @Transactional
+    public Mark create(Mark mark, MediaFile cover) {
+        return setPhotoAndSave(mark, cover);
+    }
 
-    Mark update(Mark mark, MediaFile cover);
+    @Transactional
+    public Mark update(Mark mark) {
+        return update(mark, null);
+    }
 
-    void deleteById(Long id);
+    @Transactional
+    public Mark update(Mark mark, MediaFile cover) {
+        return setPhotoAndSave(mark, cover);
+    }
 
+    @NonNull
+    private Mark setPhotoAndSave(Mark mark, MediaFile cover) {
+        if (cover != null) {
+            mark.setCover(cover);
+        }
+        Mark savedMark = repository.save(mark);
+        if (savedMark.getCover() != null) {
+            eventPublisher.publishEvent(new MarkCreatedEvent(this, savedMark.getId(), savedMark.getCover().getId(), savedMark.getCover().getOriginalFilename()));
+        }
+        return savedMark;
+    }
+
+    public void deleteById(Long id) {
+        repository.deleteById(id);
+    }
 }
